Refactor queries
